Added Placeholder Dialogue
Added an exposition dialogue for the existing characters so far in the scene (the player, a.k.a Ronin, and the "bar owner")

For MVP purposes. We can build on from this or change it. We still need to change the character sprites ofcourse and figure out more dialogue timelines.